Created Model, Migration, and Seeds for Employee table
